Rename snippets
Rename snippets to bring more clarity into their functional purpose.
