+++ src/logsearch-config/src/logstash-filters/snippets/platform-haproxy.conf
