+++ src/logsearch-config/src/logstash-filters/snippets/platform-uaa.conf
