+++ src/logsearch-config/src/logstash-filters/snippets/platform-vcap.conf
