+++ src/logsearch-config/test/logstash-filters/filter_test_helpers.rb
