+++ src/logsearch-config/test/logstash-filters/it/app-it-spec.rb
@@ -1,5 +1,5 @@
 # encoding: utf-8
-require 'test/filter_test_helpers'
+require 'test/logstash-filters/filter_test_helpers'
 require 'test/logstash-filters/it_app_helper' # app it util
 
 describe "App logs IT" do
