+++ src/logsearch-config/test/logstash-filters/it/platform-it-spec.rb
@@ -1,5 +1,5 @@
 # encoding: utf-8
-require 'test/filter_test_helpers'
+require 'test/logstash-filters/filter_test_helpers'
 require 'test/logstash-filters/it_platform_helper' # platform it util
 
 describe "Platform logs IT" do
