+++ src/logsearch-config/test/logstash-filters/snippets/platform-haproxy-spec.rb
@@ -1,12 +1,12 @@
 # encoding: utf-8
-require 'test/filter_test_helpers'
+require 'test/logstash-filters/filter_test_helpers'
 
-describe "haproxy.conf" do
+describe "platform-haproxy.conf" do
 
   before(:all) do
     load_filters <<-CONFIG
       filter {
-        #{File.read("src/logstash-filters/snippets/haproxy.conf")}
+        #{File.read("src/logstash-filters/snippets/platform-haproxy.conf")}
       }
     CONFIG
   end
