+++ src/logsearch-config/test/logstash-filters/snippets/teardown-spec.rb
@@ -1,5 +1,5 @@
 # encoding: utf-8
-require 'test/filter_test_helpers'
+require 'test/logstash-filters/filter_test_helpers'
 
 describe "teardown.conf" do
